Changed the HTML input validation pattern, as well as the serverside form parser in the route to allow floats in precipitation data. Fixes #16

+++ make_a_data_object/routes.py
@@ -15,7 +15,7 @@ def index():
 @app.route('/make', methods=['POST'])
 def submit():
     a = request.form['abstract']
-    p = [int(p) for p in request.form['precipitation'].split(',')]
+    p = [float(p) for p in request.form['precipitation'].split(',')]
     try:
         s = int(request.form.get('smoothing'))
     except ValueError:
